ADD OPTION TO COLOUR ON CLICK
Added buttons to html to be able to choose to either colour on
click or on mouseover. CSS styling added to have buttons inline
with the rest of the user selection options.

Added eventlistener to grid cells to be able to colour on click
and added an if statement to both event listeners on grid cells to
return if not current selected drawing method.